+++ js/jquery.terminal.d.ts
@@ -776,7 +776,9 @@ interface JQueryTerminal<TElement = HTMLElement> extends JQuery<TElement> {
     next(selector?: JQuery.Selector): this;
     next(): JQueryTerminal;
     focus(handler?: JQuery.TypeEventHandler<TElement, null, TElement, TElement, 'focus'> | false): this;
-    focus(toggle?: boolean): JQueryTerminal;
+    focus(toggle?: boolean, silent?: boolean): JQueryTerminal;
+    blur(handler?: JQuery.TypeEventHandler<TElement, null, TElement, TElement, 'blur'> | false):  this;
+    blur(silent?: boolean): JQueryTerminal;
     freeze(toggle?: boolean): JQueryTerminal;
     frozen(): boolean;
     enable(silent?: boolean): JQueryTerminal;